+++ routes/web.php
@@ -37,8 +37,8 @@ Route::middleware(['auth'])->group(function () {
         Route::put('/spj/{spj}', [SpjController::class, 'update'])->name('spj.update');
         Route::delete('/spj/{spj}', [SpjController::class, 'destroy'])->name('spj.destroy');
         Route::delete('/spj/spk/{no_spk}', [SpjController::class, 'destroySpk'])->name('spj.destroy-spk')->where('no_spk', '.*');
-        Route::get('/cari-barang', [SpjController::class, 'cariBarang'])->name('cari-barang');
-        Route::get('/unduh', [SpjController::class, 'unduh'])->name('unduh');
+        Route::get('/cari-barang', [SpjController::class, 'cariBarang'])->middleware('throttle:120,1')->name('cari-barang');
+        Route::get('/unduh', [SpjController::class, 'unduh'])->middleware('throttle:30,1')->name('unduh');
 
         Route::get('/input-realisasi/pilih-bulan', [InputRealisasiController::class, 'pilihBulan'])->name('input-realisasi.pilih-bulan');
         Route::get('/input-realisasi', [InputRealisasiController::class, 'index'])->name('input-realisasi.index');
@@ -49,31 +49,35 @@ Route::middleware(['auth'])->group(function () {
         Route::post('/input-realisasi/kirim-laporan', [InputRealisasiController::class, 'kirimLaporan'])->name('input-realisasi.kirim-laporan');
 
         Route::get('/realisasi', [RealisasiController::class, 'index'])->name('realisasi.index');
-        Route::get('/realisasi/unduh', [RealisasiController::class, 'unduh'])->name('realisasi.unduh');
+        Route::get('/realisasi/unduh', [RealisasiController::class, 'unduh'])->middleware('throttle:30,1')->name('realisasi.unduh');
         Route::get('/cetak', [CetakController::class, 'show'])->name('cetak');
-        Route::post('/cetak/check', [CetakController::class, 'check'])->name('cetak.check');
-        Route::get('/cetak/unduh', [CetakController::class, 'unduh'])->name('cetak.unduh');
-        Route::post('/kunci-laporan/toggle', [KunciLaporanController::class, 'toggle'])->name('kunci-laporan.toggle');
-        Route::post('/kunci-laporan/status', [KunciLaporanController::class, 'updateStatus'])->name('kunci-laporan.status');
+        Route::post('/cetak/check', [CetakController::class, 'check'])->middleware('throttle:60,1')->name('cetak.check');
+        Route::get('/cetak/unduh', [CetakController::class, 'unduh'])->middleware('throttle:30,1')->name('cetak.unduh');
+        Route::post('/kunci-laporan/toggle', [KunciLaporanController::class, 'toggle'])
+            ->middleware('permission:kunci-laporan-bm')
+            ->name('kunci-laporan.toggle');
+        Route::post('/kunci-laporan/status', [KunciLaporanController::class, 'updateStatus'])
+            ->middleware('permission:verifikasi-laporan-bm')
+            ->name('kunci-laporan.status');
 
         Route::get('/acuan', [AcuanController::class, 'index'])->name('acuan.index');
         Route::post('/acuan', [AcuanController::class, 'store'])->name('acuan.store');
-        Route::post('/acuan/import', [AcuanController::class, 'import'])->name('acuan.import');
-        Route::post('/acuan/destroy-all', [AcuanController::class, 'destroyAll'])->name('acuan.destroy-all');
+        Route::post('/acuan/import', [AcuanController::class, 'import'])->middleware('throttle:10,1')->name('acuan.import');
+        Route::post('/acuan/destroy-all', [AcuanController::class, 'destroyAll'])->middleware('throttle:10,1')->name('acuan.destroy-all');
         Route::delete('/acuan/{acuan}', [AcuanController::class, 'destroy'])->name('acuan.destroy');
 
         Route::get('/rekapan', [RekapanController::class, 'index'])->name('rekapan.index');
     });
 
-    Route::prefix('admin')->name('admin.')->group(function () {
-        Route::get('/user', [UserController::class, 'index'])->name('user.index');
-        Route::post('/user', [UserController::class, 'store'])->name('user.store');
-        Route::put('/user/{user}', [UserController::class, 'update'])->name('user.update');
-        Route::delete('/user/{user}', [UserController::class, 'destroy'])->name('user.destroy');
+    Route::prefix('admin')->name('admin.')->middleware('role:super_admin|admin_kcd')->group(function () {
+        Route::get('/user', [UserController::class, 'index'])->middleware('permission:kelola-user')->name('user.index');
+        Route::post('/user', [UserController::class, 'store'])->middleware('permission:kelola-user')->name('user.store');
+        Route::put('/user/{user}', [UserController::class, 'update'])->middleware('permission:kelola-user')->name('user.update');
+        Route::delete('/user/{user}', [UserController::class, 'destroy'])->middleware('permission:kelola-user')->name('user.destroy');
 
         Route::get('/kode-barang', [KodeBarangController::class, 'index'])->name('kode-barang.index');
         Route::post('/kode-barang', [KodeBarangController::class, 'store'])->name('kode-barang.store');
-        Route::post('/kode-barang/import', [KodeBarangController::class, 'import'])->name('kode-barang.import');
+        Route::post('/kode-barang/import', [KodeBarangController::class, 'import'])->middleware('throttle:10,1')->name('kode-barang.import');
         Route::delete('/kode-barang/{kodeBarang}', [KodeBarangController::class, 'destroy'])->name('kode-barang.destroy');
     });
 });
